Added auth to frontend and backend

+++ main.go
@@ -1,16 +1,21 @@
 package main
 
 import (
+	"context"
 	"database/sql"
 	"encoding/json"
 	"fmt"
 	"log"
 	"net/http"
+	"os"
 	"strconv"
 	"sync"
 	"time"
 
+	"github.com/golang-jwt/jwt/v5"
+	"github.com/joho/godotenv"
 	_ "github.com/lib/pq"
+	"golang.org/x/crypto/bcrypt"
 )
 
 type Result struct {
@@ -27,13 +32,31 @@ type Website struct {
 	Status string `json:"status"`
 }
 
+type User struct {
+	ID       int    `json:"id"`
+	Email    string `json:"email"`
+	Password string `json:"-"`
+}
+
+type AuthReq struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
+
+type AuthResponse struct {
+	ID    int    `json:"id"`
+	Email string `json:"email"`
+	Token string `json:"token"`
+}
+
 var db *sql.DB
+var jwtSecret []byte
 
 func enableCORS(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Access-Control-Allow-Origin", "*")
 		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, DELETE, OPTIONS")
-		w.Header().Set("Access-Control-Allow-Headers", "Content-Type")
+		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
 
 		if r.Method == "OPTIONS" {
 			w.WriteHeader(http.StatusOK)
@@ -73,8 +96,9 @@ func checkSite(url string) Result {
 }
 
 func handleGetWebsites(w http.ResponseWriter, r *http.Request) {
+	userID := r.Context().Value("user_id").(int)
 
-	rows, err := db.Query("SELECT id, name, url, status FROM monitors")
+	rows, err := db.Query("SELECT id, name, url, status FROM monitors WHERE user_id = $1", userID)
 	if err != nil {
 		http.Error(w, "Database error", http.StatusInternalServerError)
 		return
@@ -102,14 +126,16 @@ func handleGetWebsites(w http.ResponseWriter, r *http.Request) {
 
 func handleCreateWebsite(w http.ResponseWriter, r *http.Request) {
 	var newSite Website
+	userID := r.Context().Value("user_id").(int)
+
 	err := json.NewDecoder(r.Body).Decode(&newSite)
 	if err != nil {
 		http.Error(w, "Invalid body request", http.StatusBadRequest)
 		return
 	}
 
-	err = db.QueryRow("INSERT INTO monitors (name, url, status) VALUES ($1, $2, $3) RETURNING id",
-		newSite.Name, newSite.URL, "unknown").Scan(&newSite.ID)
+	err = db.QueryRow("INSERT INTO monitors (name, url, status, user_id) VALUES ($1, $2, $3, $4) RETURNING id",
+		newSite.Name, newSite.URL, "unknown", userID).Scan(&newSite.ID)
 
 	if err != nil {
 		http.Error(w, "Database error", http.StatusInternalServerError)
@@ -123,6 +149,7 @@ func handleCreateWebsite(w http.ResponseWriter, r *http.Request) {
 func handleGetWebsiteByID(w http.ResponseWriter, r *http.Request) {
 	idStr := r.PathValue("id")
 	id, err := strconv.Atoi(idStr)
+	userID := r.Context().Value("user_id").(int)
 
 	if err != nil {
 		http.Error(w, "Invalid ID", http.StatusBadRequest)
@@ -130,7 +157,7 @@ func handleGetWebsiteByID(w http.ResponseWriter, r *http.Request) {
 	}
 
 	var site Website
-	err = db.QueryRow("SELECT id, name, url, status FROM monitors where id = $1", id).Scan(&site.ID, &site.Name, &site.URL, &site.Status)
+	err = db.QueryRow("SELECT id, name, url, status FROM monitors where id = $1 AND user_id = $2", id, userID).Scan(&site.ID, &site.Name, &site.URL, &site.Status)
 
 	if err == sql.ErrNoRows {
 		http.Error(w, "Website not found", http.StatusNotFound)
@@ -147,13 +174,14 @@ func handleGetWebsiteByID(w http.ResponseWriter, r *http.Request) {
 func handleDeleteWebsite(w http.ResponseWriter, r *http.Request) {
 	idStr := r.PathValue("id")
 	id, err := strconv.Atoi(idStr)
+	userID := r.Context().Value("user_id").(int)
 
 	if err != nil {
 		http.Error(w, "Invalid ID", http.StatusBadRequest)
 		return
 	}
 
-	result, err := db.Exec("DELETE FROM monitors WHERE id = $1", id)
+	result, err := db.Exec("DELETE FROM monitors WHERE id = $1 AND user_id = $2", id, userID)
 	if err != nil {
 		http.Error(w, "Database error", http.StatusInternalServerError)
 		return
@@ -166,6 +194,127 @@ func handleDeleteWebsite(w http.ResponseWriter, r *http.Request) {
 	}
 
 	w.WriteHeader(http.StatusNoContent)
+}
+
+func handleRegister(w http.ResponseWriter, r *http.Request) {
+	var req AuthReq
+
+	err := json.NewDecoder(r.Body).Decode(&req)
+
+	if err != nil {
+		http.Error(w, "Invalid body request", http.StatusBadRequest)
+		return
+	}
+
+	if req.Email == "" || req.Password == "" {
+		http.Error(w, "Invalid body request", http.StatusBadRequest)
+		return
+	}
+
+	hashedBytes, err := bcrypt.GenerateFromPassword([]byte(req.Password), bcrypt.DefaultCost)
+	if err != nil {
+		http.Error(w, "Failed to hash password", http.StatusInternalServerError)
+		return
+	}
+	passwordHash := string(hashedBytes)
+
+	var userID int
+
+	err = db.QueryRow("INSERT INTO users (email, password_hash) VALUES ($1, $2) RETURNING id", req.Email, passwordHash).Scan(&userID)
+
+	if err != nil {
+		http.Error(w, "Email already registered", http.StatusBadRequest)
+		return
+	}
+
+	token, err := createToken(userID)
+	if err != nil {
+		http.Error(w, "Failed to generate token", http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusCreated)
+	json.NewEncoder(w).Encode(AuthResponse{ID: userID, Email: req.Email, Token: token})
+}
+
+func handleLogin(w http.ResponseWriter, r *http.Request) {
+	var req AuthReq
+
+	err := json.NewDecoder(r.Body).Decode(&req)
+
+	if err != nil {
+		http.Error(w, "Invalid body request", http.StatusBadRequest)
+		return
+	}
+
+	var userID int
+	var hashedPw string
+
+	err = db.QueryRow("SELECT id, password_hash FROM users WHERE email = $1", req.Email).Scan(&userID, &hashedPw)
+
+	if err == sql.ErrNoRows {
+		http.Error(w, "Invalid email or password", http.StatusUnauthorized)
+		return
+	}
+
+	err = bcrypt.CompareHashAndPassword([]byte(hashedPw), []byte(req.Password))
+	if err != nil {
+		http.Error(w, "Invalid email or password", http.StatusUnauthorized)
+		return
+	}
+
+	token, err := createToken(userID)
+	if err != nil {
+		http.Error(w, "Failed to generate token", http.StatusInternalServerError)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	json.NewEncoder(w).Encode(AuthResponse{ID: userID, Email: req.Email, Token: token})
+}
+
+func createToken(userID int) (string, error) {
+	claims := jwt.MapClaims{
+		"user_id": userID,
+		"exp":     time.Now().Add(24 * time.Hour).Unix(), // expires in 24 hours
+	}
+
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
+	return token.SignedString(jwtSecret)
+}
+
+func requireAuth(next http.HandlerFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		authHeader := r.Header.Get("Authorization")
+		if authHeader == "" || len(authHeader) < 7 || authHeader[:7] != "Bearer " {
+			http.Error(w, "Unauthorized: missing token", http.StatusUnauthorized)
+			return
+		}
+
+		tokenString := authHeader[7:]
+
+		token, err := jwt.Parse(tokenString, func(t *jwt.Token) (interface{}, error) {
+			return jwtSecret, nil
+		})
+
+		if err != nil || !token.Valid {
+			http.Error(w, "Unauthorized: invalid token", http.StatusUnauthorized)
+			return
+		}
+
+		claims, ok := token.Claims.(jwt.MapClaims)
+		if !ok {
+			http.Error(w, "Unauthorized", http.StatusUnauthorized)
+			return
+		}
+
+		userID := int(claims["user_id"].(float64))
+
+		ctx := context.WithValue(r.Context(), "user_id", userID)
+		next(w, r.WithContext(ctx))
+	}
 }
 
 func checkAllMonitors() {
@@ -221,7 +370,12 @@ func startBackgroundChecker() {
 }
 
 func main() {
-	connStr := "postgres://localhost:5432/uptime?sslmode=disable"
+	_ = godotenv.Load()
+
+	secret := os.Getenv("JWT_SECRET")
+	jwtSecret = []byte(secret)
+
+	connStr := os.Getenv("DATABASE_URL")
 	mux := http.DefaultServeMux
 
 	var err error
@@ -237,10 +391,12 @@ func main() {
 	}
 	fmt.Println("Connected to PostgreSQL successfully!")
 
-	http.HandleFunc("GET /websites", handleGetWebsites)
-	http.HandleFunc("GET /websites/{id}", handleGetWebsiteByID)
-	http.HandleFunc("POST /websites", handleCreateWebsite)
-	http.HandleFunc("DELETE /websites/{id}", handleDeleteWebsite)
+	http.HandleFunc("GET /websites", requireAuth(handleGetWebsites))
+	http.HandleFunc("GET /websites/{id}", requireAuth(handleGetWebsiteByID))
+	http.HandleFunc("POST /websites", requireAuth(handleCreateWebsite))
+	http.HandleFunc("DELETE /websites/{id}", requireAuth(handleDeleteWebsite))
+	http.HandleFunc("POST /auth/register", handleRegister)
+	http.HandleFunc("POST /auth/login", handleLogin)
 
 	go startBackgroundChecker()
